Clear canvas before redrawing
When enabling and disabling the circumcircle option it can happen that
parts of the circumcircles remain on the canvas. That happens because
not every bit of it is redrawn.
With this change we clear the canbas before performing any draw
operations and, hence, start with a pristine state without any
remainders such as circumcircles.